# New: Added ignore option (fixes #115)