--- a/glob.js
+++ b/glob.js
@@ -56,6 +56,7 @@ var setopts = common.setopts
 var ownProp = common.ownProp
 var inflight = require('inflight')
 var util = require('util')
+var isIgnored = common.isIgnored
 
 var once = require('once')
 
@@ -270,13 +271,16 @@ Glob.prototype._process = function (pattern, index, inGlobStar, cb) {
 
   var abs = this._makeAbs(read)
 
+  //if ignored, skip _processing
+  if (isIgnored(this, read, true))
+    return cb()
+
   var isGlobStar = remain[0] === minimatch.GLOBSTAR
   if (isGlobStar)
     this._processGlobStar(prefix, read, abs, remain, index, inGlobStar, cb)
   else
     this._processReaddir(prefix, read, abs, remain, index, inGlobStar, cb)
 }
-
 
 Glob.prototype._processReaddir = function (prefix, read, abs, remain, index, inGlobStar, cb) {
   var self = this