--- a/test/bash-results.json
+++ b/test/bash-results.json
@@ -101,6 +101,8 @@
     "./test/global-leakage.js",
     "./test/globstar-match.js",
     "./test/has-magic.js",
+    "./test/ignore-sync.js",
+    "./test/ignore.js",
     "./test/mark-sync.js",
     "./test/mark.js",
     "./test/negation-test.js",