fixes #115 

In addition to ```minimatch``` check, the code checks if a particular pattern ends with a glob star ```**```, if it does, adds the directory to the ignore list and stops traversal for the directory matching the pattern. 

Also, if, for example, ```a/**``` is added to the ignore list, it removes ```a``` from the resulting match list.