@UltCombo 2015-01-02
Odd how the only failing test is due to 2 paths out of order in Node 0.11 only.

Nice work, I'll start experimenting with this tomorrow hopefully.

//cc @contra @sindresorhus