@vegetableman 2015-01-02
@UltCombo had those failed cases randomly thrown on my local as well but they may be due to race conditions mentioned in the README. Travis is clear now.