@UltCombo 2015-01-08
Sorry, hadn't had time to test it throughout. I just pushed a [new branch](https://github.com/sindresorhus/globby/tree/glob-ignore) to globby using node-glob's `ignore` option in the async API.

Ignoring some files in a directory through the `ignore` option is about twice slower than globby, looks like it is because you're calling the minimatch function [inside a loop](https://github.com/isaacs/node-glob/pull/151/files#diff-d7716810d825f4b55d18727c3ccb24e6R192) which parses the pattern again in each iteration. Ideally, only a single [Minimatch class](https://github.com/isaacs/minimatch#minimatch-class) instance should be created for each `ignore` pattern. See how [globby handles it currently](https://github.com/sindresorhus/globby/blob/master/index.js#L34-L36).

Ignoring a whole dir (`dir/**`) is throwing a `RangeError: Maximum call stack size exceeded`, this might be a problem with the benchmark though.