@contra 2015-01-08
@UltCombo thanks for taking the time to benchmark it :+1: 

So I put $100 on bountysource for this but it was on the other issue #, not sure how to move it over here but I'll contact them about that.