@vegetableman 2015-01-10
@UltCombo Thanks for the benchmarks.

> you're calling the minimatch function inside a loop which parses the pattern again in each iteration.Ideally, only a single Minimatch class instance should be created for each ignore pattern. See how globby handles it currently.

Creating only one instance now.

> Ignoring a whole dir (dir/**) is throwing a RangeError: Maximum call stack size exceeded,

Using setImmediate to return the callback has fixed the issue. Ref:- https://github.com/logicalparadox/matcha/issues/9

And following are the results from the existing globby benchmarks:
```
matcha bench

                      negative globs (some files inside dir)
             229 op/s » globby async (working directory)
             223 op/s » globby async (upstream/master)
             241 op/s » globby sync (working directory)
             226 op/s » globby sync (upstream/master)
              59 op/s » glob-stream

                      negative globs (whole dir)
          10,075 op/s » globby async (working directory)
             368 op/s » globby async (upstream/master)
             236 op/s » globby sync (working directory)
             235 op/s » globby sync (upstream/master)
              66 op/s » glob-stream

                      multiple positive globs
              86 op/s » globby async (working directory)
              89 op/s » globby async (upstream/master)
              77 op/s » globby sync (working directory)
              74 op/s » globby sync (upstream/master)
              36 op/s » glob-stream


  Suites:  3
  Benches: 15
  Elapsed: 26,361.82 ms
```