@UltCombo 2015-01-10
Nice!
Btw, so far I've only updated the async logic to use the `ignore` option, so you can ignore the sync results for now.