@vegetableman 2015-01-10
Updated benchmarks on fixing globby sync:-

```
matcha bench

                      negative globs (some files inside dir)
             223 op/s » globby async (working directory)
             222 op/s » globby async (upstream/master)
             432 op/s » globby sync (working directory)
             254 op/s » globby sync (upstream/master)
              68 op/s » glob-stream

                      negative globs (whole dir)
          11,760 op/s » globby async (working directory)
             409 op/s » globby async (upstream/master)
          18,977 op/s » globby sync (working directory)
             244 op/s » globby sync (upstream/master)
              73 op/s » glob-stream

                      multiple positive globs
              98 op/s » globby async (working directory)
              96 op/s » globby async (upstream/master)
             516 op/s » globby sync (working directory)
              81 op/s » globby sync (upstream/master)
              37 op/s » glob-stream


  Suites:  3
  Benches: 15
  Elapsed: 23,045.84 ms
```