@UltCombo 2015-01-10
Oh, I was about to start working on that, thank you very much!
Mind sending a PR to globby later?