@isaacs 2015-01-11
@vegetableman The tests are thorough, and it looks like you've got the API pretty much perfect and working.  Good job!

However, there's a **lot** of stuff in this patch that is not relevant to the change.  I started commenting on it, but gave up after a while.

1. Single quotes have been changed to double.  Including in words like `isn't` in comments, even.  Don't do that.  This project uses single-quotes exclusively for strings, as of 82b8568.
2. The if/else formatting has been changed in many places.  There are two ways that this project does if blocks.  If it is a single-line, it should not use braces.  If it is a multi-line, it should use braces of the form of `} else {` all on one line.
3. Spacing around conditionals and function calls.  There should be spaces between the `if` and the opening `(` of the conditional.  Function definitions should have a space between the name and the opening `(`.
4. Many unnecessary `setImmediate` calls are added, I'm not sure why.  They aren't relevant to this feature.  If you are finding cases with strange timing errors, please provide example code and post a new issue about that.

These things make it much harder than necessary to review the change.

In general, a commit for a feature should be as small as possible, and no smaller; it should do exactly the feature that is being added, and nothing else; it should match the style of the surrounding code so well that no one can tell that the original author didn't write it themselves.