@UltCombo 2015-01-11
> Have sent fix for sync. Matcha fix? Do you mean the setImmediate fix? It's part of node-glob.

Seeing as this seems to be a [Matcha issue](https://github.com/logicalparadox/matcha/issues/9) (as you've linked previously), I believe it would make sense to workaround the issue in the benchmark tests instead of the lib core.

That is, add `setImmediate` to call the benchmark test's callback instead of polluting the node-glob core with it.