@vegetableman 2015-01-11
@isaacs Thanks for reviewing it. Have fixed the formatting and quote issues. Sorry for that. Removed ```setImmediate``` calls which were added to fix call stack errors on globby benchmarks/test issues. 

> What if the pattern ends with multiple /** parts? Eg, foo/**/**/** is functionally equivalent to foo/**

 Is it supposed to behave differently?. right now, fetching ```foo/**``` and ```foo/**/**/**``` return the same number of results.