@isaacs 2015-01-11
Note: Travis error is spurious, fixed on aaffe249119c4640ed71e186d40516aa8e7e2d4e.

> Is it supposed to behave differently?. right now, fetching `foo/**` and `foo/**/**/**` return the same number of results.

No, it's supposed to behave the same, but the issue is that you're doing special treatment for things that end in `/**`, and stripping it off.  Why not `.replace(/(\/\*\*)+$/, '')` instead, to strip off all trailing `/**` segments?