@vegetableman 2015-01-11
@isaacs Thanks for testing. Took another crack at it, have fixed the pattern matching issues, i believe once and for all. All the above issues has been addressed.

Would be great if you consider/test these changes once, before I revert back.

> It's a premature optimization.

I don't think so. It's important because let's say i ignore node_modules folder using ```node_modules/**```, then the folder won't be read at all saving huge processing time. Please refer the above benchmark results. 

With the change:-
```
           negative globs (whole dir)
           7,233 op/s » globby async (working directory) // with the change
             409 op/s » globby async (upstream/master)
          16,077 op/s » globby sync (working directory) // with the change
             246 op/s » globby sync (upstream/master)
              71 op/s » glob-stream
```

without it:-
```
             negative globs (whole dir)
             231 op/s » globby async (working directory) // without the change
             379 op/s » globby async (upstream/master)
             428 op/s » globby sync (working directory)  // without the change
             247 op/s » globby sync (upstream/master)
              73 op/s » glob-stream
```